+++ packages/titus-frontend/.eslintrc
@@ -10,7 +10,6 @@
     "browser": true
   },
   "rules": {
-    "prettier/prettier": "error",
     "import/order": [
       "error",
       {
