+++ packages/titus-frontend/src/components/auth-providers/utils/schemas.js
@@ -1,7 +1,6 @@
 import * as yup from 'yup'
-
-import { AUTH_PROVIDERS } from '../../../constants'
-import i18n from '../../../i18n'
+import { AUTH_PROVIDERS } from 'lib/constants'
+import i18n from 'lib/i18n'
 
 let password
 if (process.env.REACT_APP_AUTH_PROVIDER === AUTH_PROVIDERS.MEM) {
