+++ packages/titus-frontend/src/components/dashboard/dashboard.mdx
@@ -1,5 +1,5 @@
 import { Story } from '@storybook/addon-docs/blocks'
-import { StyledPreview as Preview } from '../../../.storybook/decorator-components'
+import { StyledPreview as Preview } from '.storybook/decorator-components'
 import Dashboard from './'
 
 ## Dashboard
@@ -7,5 +7,5 @@ import Dashboard from './'
 Dashboard component.
 
 <Preview>
-  <Story id='dashboard--default'/>
+  <Story id="dashboard--default" />
 </Preview>
