+++ packages/titus-frontend/src/components/dashboard/dashboard.story.js
@@ -1,14 +1,11 @@
 import React from 'react'
 
-import Layout from '../layout'
-
 import docs from './dashboard.mdx'
 
 import Dashboard from './'
 
 export default {
   title: 'Dashboard',
-  decorators: [story => <Layout>{story()}</Layout>],
   parameters: {
     docs: {
       page: docs
