+++ packages/titus-frontend/src/components/loading/loading.mdx
@@ -1,10 +1,10 @@
 import { Story } from '@storybook/addon-docs/blocks'
-import { StyledPreview as Preview } from '../../../.storybook/decorator-components'
+import { StyledPreview as Preview } from '.storybook/decorator-components'
 
 ## Loader
 
 Loader component for project.
 
 <Preview>
-  <Story id='loader--default'/>
+  <Story id="loader--default" />
 </Preview>
