+++ packages/titus-frontend/src/components/loading/loading.story.js
@@ -1,14 +1,11 @@
 import React from 'react'
 
-import Layout from '../layout'
-
 import docs from './loading.mdx'
 
 import Loader from './'
 
 export default {
   title: 'Loader',
-  decorators: [story => <Layout>{story()}</Layout>],
   parameters: {
     docs: {
       page: docs
