+++ packages/titus-frontend/src/components/login-form/login-form.story.js
@@ -1,14 +1,11 @@
 import React from 'react'
 
-import Layout from '../layout'
-
 import docs from './login-form.mdx'
 
 import LoginForm from './'
 
 export default {
   title: 'Login Form',
-  decorators: [story => <Layout>{story()}</Layout>],
   parameters: {
     docs: {
       page: docs
