+++ packages/titus-frontend/src/components/logo/logo.mdx
@@ -1,10 +1,10 @@
 import { Story } from '@storybook/addon-docs/blocks'
-import { StyledPreview as Preview } from '../../../.storybook/decorator-components'
+import { StyledPreview as Preview } from '.storybook/decorator-components'
 
 ## Logo
 
 Logo component for project.
 
 <Preview>
-  <Story id='logo--default'/>
+  <Story id="logo--default" />
 </Preview>
