+++ packages/titus-frontend/src/components/user-info/index.js
@@ -1,7 +1,6 @@
 import React, { useState, useContext, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-
-import { AuthContext } from '../authentication/authentication-context'
+import { AuthContext } from 'components/authentication/authentication-context'
 
 // This component was created only to test the auth endpoint
 const UserInfo = () => {
