--- packages/titus-frontend/src/history.js
@@ -1,2 +1,0 @@
-import { createBrowserHistory } from 'history'
-export default createBrowserHistory()
