+++ packages/titus-frontend/src/lib/i18n.js
@@ -1,9 +1,8 @@
 import i18n from 'i18next'
 import detector from 'i18next-browser-languagedetector'
 import { initReactI18next } from 'react-i18next'
-
-import locale from './locale'
-import { LANGUAGES } from './constants'
+import locale from 'lib/locale'
+import { LANGUAGES } from 'lib/constants'
 
 const resources = {
   en: {
