+++ packages/titus-frontend/src/lib/pino.test.js
@@ -1,4 +1,5 @@
 import pino, { log } from './pino'
+
 let warnMessage
 let infoMessage
